+++ src/hexmaze/__init__.py
@@ -1,0 +1,8 @@
+from . import core, utils, barrier_shift, plotting
+
+from .core import *
+from .utils import *
+from .barrier_shift import *
+from .plotting import *
+
+__all__ = core.__all__ + utils.__all__ + barrier_shift.__all__ + plotting.__all__
